[Action] Ensure `pylint` usage errors are caught
Debugging the pervious commit highlighted that we were passing the
action if there was a usage error in the invocation of `pylint`.

This method seems gnarly, but seems to be the only way to my noob
JS explorations (it's been a _long_ time).

Whilst debugging why catching this error wasn't working I noticed that
all of the tests for `actions/exec` use the `.catch` approach over a
`try` block.

Signed-off-by: Tom Cowland <tom@foundry.com>